# Conflicts:
#	lib/features/data_usage/presentation/pages/data_usage_view.dart

+++ lib/features/active_package_details/domain/entity/active_packages_details_entity.dart
@@ -23,7 +23,7 @@ class ActivePackagesDetailsEntity {
   final DateTime activeUntil;
   final double renewalFee;
   final int totalPackageCount;
-  final int availableVolumeGb;
+  final double availableVolumeGb;
   final int totalVolumeGb;
   final List<AdOnEntity> activeAddOns;
