+++ lib/presentation/page_component/package_info_card.dart
@@ -160,7 +160,7 @@ class PackageInfoCard extends StatelessWidget {
 }
 
 class _DataUsageBar extends StatelessWidget {
-  final int usedGB;
+  final double usedGB;
   final int totalGB;
 
   const _DataUsageBar({required this.usedGB, required this.totalGB});
